tests/wpt/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssUnparsedValue.html

<!doctype html>
<meta charset="utf-8">
<title>CSSUnparsedValue</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#unparsedvalue-objects">
<meta name="assert" content="Test CSSUnparsedValue constructor and members">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<body>
<div id="log">
<script>
'use strict';

const gTestArguments = [
  {
    description: 'no arguments',
    args: [],
  },
  {
    description: 'an empty string',
    args: [''],
  },
  {
    description: 'a CSSVariableReferenceValue',
    args: [new CSSVariableReferenceValue('--foo')],
  },
  {
    description: 'mix of strings and CSSVariableReferenceValues',
    args: [
      'foo',
      'bar',
      new CSSVariableReferenceValue('--A'),
      'baz',
      new CSSVariableReferenceValue('--B')
    ],
  },
];

for (const {args, description} of gTestArguments) {
  test(() => {
    const result = new CSSUnparsedValue(args);

    assert_not_equals(result, null,
        'A CSSUnparsedValue should be created');
    assert_array_equals(result, args,
        'Content of CSSUnparsedValue should be same as the arguments ' +
        'passed in the constructor');
  }, 'CSSUnparsedValue can be constructed from ' + description);
}

test(() => {
  let result = new CSSUnparsedValue([new CSSVariableReferenceValue('--foo')]);

  result[0] = 'A';
  assert_equals(result[0], 'A', 'Item should be updated to new value');
}, 'Can update item in CSSUnparsedValue to a string');

test(() => {
  let result = new CSSUnparsedValue(['foo']);

  result[0] = new CSSVariableReferenceValue('--A');
  assert_style_value_equals(result[0], new CSSVariableReferenceValue('--A'),
      'Item should be updated to new value');
}, 'Can update item in CSSUnparsedValue to a variable reference');

test(() => {
  let result = new CSSUnparsedValue([]);

  result[0] = new CSSVariableReferenceValue('--A');
  assert_equals(result.length, 1,
      'Length of CSSUnparsedValue should have increased');
  assert_style_value_equals(result[0], new CSSVariableReferenceValue('--A'),
      'New item should be appended');

  result[1] = 'foo';
  assert_equals(result.length, 2,
      'Length of CSSUnparsedValue should have increased');
  assert_equals(result[1], 'foo', 'New item should be appended');
}, 'Can append items to CSSUnparsedValue');

test(() => {
  const result = new CSSUnparsedValue(['foo', 'bar']);
  assert_equals(result[3], undefined);
}, 'Getting invalid index in CSSUnparsedValue returns undefined');

test(() => {
  let result = new CSSUnparsedValue(['foo', 'bar']);
  assert_throws(new RangeError(), () => result[3] = 'foo');
}, 'Setting invalid index in CSSUnparsedValue throws RangeError');

</script>
