LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssUnparsedValue-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSVariableReferenceValue

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSVariableReferenceValue
